Keystore 2.0: Fix misinterpretation of the versionNumber.
Keystore 2.0 assumed that the version number in the HWInfo struct
corresponds to the HAL version. This is actually not the case as per the
KeyMint spec. The legacy wrapper sets the version number as expected but
genuine Keymint implementations might not. This match ignores the
versionNumber filed of genuine KeyMint implementations and replaces it
with the HAL version.

Bug: 187862706
Test: N/A
Merged-In: I392d4ec3c00f22f11920abe44cc3375b7266508e
Change-Id: I392d4ec3c00f22f11920abe44cc3375b7266508e