Add (fake) CompOS key generation.
Note: the CompOS work here is all still behind an if (false).

Added a new class, FakeCompOs, to allow prototyping of the interface
and implementation of the key management work that will be in CompOS.

Extensive refactoring of the certificate generation code to support
both a self-signed cert and our certificate for the CompOS key.

Bug: 190166662
Test: presubmits
Test: manual - certificate gets generated on first boot
Test: manual - certificate verifies ok on second boot
Test: manual inspection of the generated certs' text form
Change-Id: I2f2f043427774c0805e963dfe582feb8d3eac3a4