Replace FakeCompOs with real CompOs
Change the implementation of verifyCompOsKey to execute the
compos_verify_key binary to verify the pending and/or current
instance.

Then delete FakeCompOs since there is no remaiing usage of it.

Note that this CL also removes the "if (false)" that has been guarding
this code. Instead it is active if we have a userdebug or eng build,
kvm is enabled, and the CompOs APEX is present. Currently this will be
very few devices; even on those nothing significant will happen unless
valid CompOs key files and artifacts are present which can only occur
via test tools for now.

The flag is retained to give an easy kill switch.

Bug: 193603140
Test: In progress
Change-Id: Id9de0852fafc54a825a04b1bee2637a7965214e6