Fix checking of compos.info
We now load the file and its signature before odrefresh deletes them
rather than after.

And remove the check that all files listed are still
present. odrefresh can and does delete some if it decides they are not
needed, and that's fine. odrefresh will also complain if files that
should be present are not, so we will still detect if an attacker has
deleted artifacts.

Bug: 161471326
Bug: 209572241
Test: composd_cmd staged-apex-compile, reboot, odsign accepts artifacts
Change-Id: I449bd19a691c6070ce5b7295a5e4c8a7a61edfbd