Add diced.microdroid
diced.microdroid is diced built for microdroid. The only difference is
that it's built with `bootstrap: true` so that it can start early when
APEXes are not activated.

Previously, diced was started after APEXes are activated. However, it
actually has to be started before because microdroid_manager depends on
diced to get the per-VM secret that it uses to decrypt the instance disk
where the list of APEXes to be activated is stored.

Bug: 214231981
Test: run microdroid
Change-Id: I74767667b0fa3a5c83c3b62f591b7f5a862791f7