Avoid useless recompile
In certain cases where CompOS was present we were setting
odrefresh_status unconditionally to kCompilationRequired, meaning we
will run "odrefresh --compile", when we should set it whatever
"odrefresh --check-artifacts" returns (which might be kOkay, in which
case we can skip the compilation.

This makes little difference in practice, but it does mean a needless
divergence between the CompOS and non-CompOS cases.

Bug: 217365237
Test: Manual (see bug)
Ignore-AOSP-First: Internal has diverged due to a security fix
Change-Id: Idbd48594bb16065b9d9ffa7ecd40193aa61c19c6