Revert^2 "Fix encoding of PubKeyEd25519"
Fix the related tests which use (incorrect) hard-coded CBOR.

Add the tests to TEST_MAPPING, so that they run on presubmit when we
make any changes. (I've omitted diced_client_test - it only works on a
system where IDiceNode is present, and we don't currently have that in
the host. It's not currently run anywhere else for the same reason.)

This reverts commit d331f074467d7cd91b1e430b39f65ca4e68308de.

Bug: 233338216
Test: croot system/security/diced; atest -p
Change-Id: I7f66445baab44757e9d2f429460caa280ed16001