Refactor: disambiguate fs-verity helper functions
The original code mixes 3 operations (enable, measure, and enabling
status) in the helper functions. Refactoring to match the underlying
syscalls.

Note that the "original measureFsVerity" returns the status of fs-verity
for the file by measuring the file, which involves a hash operation.
This change also replaces it with a query to the file attribute.

Bug: 258061812
Test: odsign_e2e_tests
Change-Id: Iae3994351dc6a92fc8c299d942b6e95f8ed12ef3