keystore: remove unused Keystore1 files
Remove files from the "keystore" directory that are unused, i.e. not
referenced by an Android.bp, or in the case of header files not included
from anywhere.  This mostly includes files belonging to Keystore1 that
were missed when Keystore1 was deleted.  This also includes a couple
outdated test scripts that were never actually wired up to anything.

Bug: 171305684
Test: m; mmm system/security
Change-Id: Ie31f773b5f15b0a0f95dffa9cad109fbf3a84970